+++ .gitignore
@@ -1,5 +1,7 @@
 .pytest_cache/
 .idea/
+.cache/
+.build/
 node_modules/
 venv/
