+++ Deities/package.json
@@ -1,11 +1,13 @@
 {
   "scripts": {
     "babel": "babel",
-    "uglify": "uglifyjs"
+    "parcel": "parcel build"
   },
   "dependencies": {
     "babel-cli": "^6.26.0",
     "babel-preset-stage-2": "^6.24.1",
-    "uglify-es": "^3.3.9"
+    "jquery": "^3.3.1",
+    "parcel-bundler": "^1.9.4",
+    "raven-js": "^3.26.3"
   }
 }
