+++ Deities/src/USERSCRIPT.yml
@@ -1,14 +1,16 @@
 userscript:
-  name:           MyAnimeStream
-  author:         siku2
-  version:        1.7.0
-  description:    Get the most out of MyAnimeList by turning the page into a streaming site.
-  homepage:       https://siku2.github.io/
-  downloadURL:    https://mas.dokkeral.com/download
-  icon:           https://mas.dokkeral.com/images/logo
-  include:        /^http(?:s)?\:\/\/myanimelist\.net.*$/
-  noframes:       true
-  run-at:         document-start
+  name:                 MyAnimeStream
+  author:               siku2
+  version:              1.7.0
+  description:          Get the most out of your favourite anime site by turning the page into a streaming site.
+  homepage:             https://siku2.github.io/
+  downloadURL:          https://mas.dokkeral.com/download
+  icon:                 https://mas.dokkeral.com/images/logo
+  include:
+    - /^http(?:s)?\:\/\/myanimelist\.net.*$/
+    - /^http(?:s)?\:\/\/kitsu\.io.*$/
+  noframes:             true
+  run-at:               document-start
   require:
     - https://code.jquery.com/jquery-3.3.1.min.js
     - https://cdn.ravenjs.com/3.26.2/raven.min.js
@@ -16,9 +18,6 @@ userscript:
     - https://cdn.plyr.io/3.3.10/plyr.js
 
 settings:
-  merger:
-    order:
-      first:      "constants"
-      last:       "core"
+  entry:                "core.js"
   processor:
-    log_format:   "{script.fileid}: {message}"
+    log_format:         "{script.fileid}: {message}"
