+++ Deities/tools/greaser/lib/babel.py
@@ -1,12 +1,7 @@
-import shutil
-import subprocess
-
-NPM_LOCATION = shutil.which("npm")
+from . import npm
 
 DEFAULT_FLAGS = ["--minified", "--no-babelrc", "--presets=stage-2"]
-BABEL_ARGS = [NPM_LOCATION, "-s", "run", "babel", "--"]
 
 
 def compile_js(script: str) -> str:
-    args = BABEL_ARGS + DEFAULT_FLAGS
-    return subprocess.run(args, input=script, stdout=subprocess.PIPE, encoding="utf-8", check=True).stdout
+    return npm.run("babel", DEFAULT_FLAGS, script)
