+++ Deities/tools/greaser/lib/parcel.py
@@ -1,0 +1,9 @@
+from pathlib import Path
+
+from . import npm
+
+
+def bundle(entry_point: Path, build_dir: Path) -> str:
+    res = npm.run("parcel", [str(entry_point), "-d", str(build_dir), "--out-file", "_build.js"])
+    print(res)
+    return (build_dir / "_build.js").read_text("utf-8")
